requirements.txt
gradio==4.44.1
gradio_client==1.3.0
python-decouple==3.8
requests==2.32.3